requirements.txt
certifi==2023.7.22
charset-normalizer==3.2.0
idna==3.4
itchat-uos==1.5.0.dev0
pip==23.2.1
pypng==0.20220715.0
PyQRCode==1.2.1
requests==2.31.0
setuptools==68.1.0
urllib3==2.0.4
wheel==0.41.1